🎨 Palette: Move reusable UI components to ui/ directory
- Move SearchInputWithHint and KbdHint into `dashboard/src/components/ui/`
- Update component and test imports to reflect the new directory structure
