+++ dashboard/src/components/ui/KbdHint.tsx
